Added Features & Bugs to README, deleted messages are now displayed correctly. DEL button works as intended.

+++ src/store/socket/socketAction.js
@@ -193,10 +193,11 @@ const socketEmitMessageDelete = () => ({
   type: actions.SOCKET_EMIT_MESSAGE_DELETE,
 });
 
-const socketOnMessageDelete = (messageId, message) => ({
+const socketOnMessageDelete = (messageId, message, friendId) => ({
   type: actions.SOCKET_ON_MESSAGE_DELETE,
   messageId,
   message,
+  friendId,
 });
 
 const emitUserTypingState =
@@ -254,21 +255,33 @@ export const emitMessageSeenState =
   };
 
 export const emitMessageDelete =
-  (socket, messageId) => (dispatch, getState) => {
-    const activeChatId = getState().socket.isActiveChat.index;
-    const selectedChat = getState().socket.friends[activeChatId];
+  (socket, messageId, messageTimestamp) => (dispatch, getState) => {
+    const _messageTimestamp = new Date(messageTimestamp);
+    const allowedDeletion =
+      (new Date().getTime() - _messageTimestamp.getTime()) / 3600000;
+
+    if (allowedDeletion > 1)
+      return alert(
+        "ERROR: Messages sent more than an hour ago cannot be deleted!"
+      );
+
+    const activeChatIndex = getState().socket.isActiveChat.index;
+    const selectedChat = getState().socket.friends[activeChatIndex];
+    const authUserId = getState().auth.userId;
     dispatch(socketEmitMessageDelete());
     socket.emit(socketActions.messageDelete, {
       messageId: messageId,
       roomId: selectedChat.uniqueId,
+      senderId: authUserId,
+      recipientId: selectedChat.id,
     });
   };
 
 export const onMessageDelete =
-  (isDeleted, messageId, message) => (dispatch) => {
-    if (!isDeleted) return alert("ERROR: Message couldn't be deleted.");
-
-    dispatch(socketOnMessageDelete(messageId, message));
+  (isDeleted, messageId, message, friendId) => (dispatch) => {
+    if (!isDeleted || !friendId || !messageId)
+      return alert("ERROR: Message couldn't be deleted!");
+    dispatch(socketOnMessageDelete(messageId, message, friendId));
   };
 
 export const onOnlineStateChange =
